requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.8.6
django-cors-headers==3.13.0
Pillow==9.4.0
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
